// src/app/model/csv.model.ts
// Copyright 2017 Donuts Inc. All Rights Reserved.
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//     http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

import { Money} from './money.model';
import { CategorizedPremiumName } from './categorized-premium-name.model';
import { ReservedName} from './reserved-name.model';

class Action {
  operation: string;
  tags: string[];
  name: string;
  category?: string;
  price?: Money;
  futureCategory?: string;
  changeDate?: Date;
};

export class CSV {
  total: number;
  create: number;
  update: number;
  delete: number;
  token: string;
  expires?: Date;
  result?: {create: [ {Action} ], delete: [ {Action} ], update: [ {Action} ]};
};